Move suggestions autocomplete in another class
